Initial commit: split seg into subfiles, loss-specific heads, start working on res-net blocks
