+++ nitorch/nn/modules/__init__.py
@@ -8,6 +8,7 @@ from . import pool
 from . import registration
 from . import spatial
 from . import segmentation
+from . import seg_advanced
 from . import segreg
 
 from .base import *
@@ -18,4 +19,5 @@ from .pool import *
 from .registration import *
 from .spatial import *
 from .segmentation import *
+from .seg_advanced import *
 from .segreg import *
